refactor(spi): split module into master and pads, add Instance trait
